requirements.txt
Flask>=2.2.0
Werkzeug>=2.3.0
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.17.1
